backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
faster-whisper
google-generativeai>=0.5.4
moviepy==1.0.3
opencv-python
imageio[ffmpeg]
Pillow
numpy